+++ api/api/spec/changes.md
@@ -131,6 +131,15 @@ This agent will use **any** as IP.
 ### PUT     /agents/{agent_id}/upgrade
 * Changed parameter type **force** in request body from integer to boolean.
 
+### PUT     /agents/upgrade
+* New endpoint. Upgrade all agents or a list of them to a different Wazuh version.
+
+### PUT     /agents/upgrade_custom
+* New endpoint. Upgrade all agents or a list of them to a different Wazuh version using a local WPK file.
+
+### GET     /agents/upgrade
+* New endpoint.  Get upgrade results from all agents or a list of them.
+
 ### PUT     /agents/{agent_name}
 * Endpoint removed. Use `POST /agents/insert/quick?agent_name=name`.